requirements.txt
# Core LLM & Agent
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-genai
langgraph

# Web Framework
fastapi
uvicorn[standard]
pydantic

# Streamlit UI
streamlit
sseclient-py
requests
httpx

# Utilities
python-dotenv

# Package install
-e .